Fix removal of IME target.
In the case that the IME target is removed while it is still
the IME target we will trigger a layer assignment before
updating WindowManagerService#mInputMethodTarget. In general in
the new hierarchy model we assume that containers always have SurfaceControl's
but this is only true while they are attached to the hierarchy. Since the
IME target may stick around otherwise until it's cleared we need
to check if it has a surface before using it in this way.

Bug: 69669036
Test: go/wm-smoke
Change-Id: I296549aba45368ee5e88f6e1224024e6105207dd

Former-commit-id: 1cafdf8b989b965e201db5839a658768173145ff