Add carrier config to always play hold tone when remotely held.
By default, a remote hold tone is played when a call is remotely held.
In other words if the callee puts the caller on hold, they will hear a
tone indicating they're on hold.  Current behavior is to only do this
when the audio direction for the call is inactive. Some carriers wish to
have this behavior when the audio direction is active.
This new carrier config option is OFF by default and will be selectively
enabled for carriers which wish to play a hold tone when the audio is
active.

Test: Manual
Bug: 64099717
Change-Id: I747a5fb650c8e83a61b6512375e43e45a5ea0362

Former-commit-id: 3088942b9eb5f121a2718517b95d59f35b00205f