Wifi: Pass the multicast wakelock TAG in release
In current implementation, when the same UID allocates, and acquires
multiple wiFi multicast wakelocks, the tag associated with the lock is
passed in the acquire method, and stored in the multicasters list.
However, when the UID tries to release the lock, the tag is not included
in the method call, and all the wakelocks found in the list of
multicasters for this UID are released.
This behavior does not reflect the expected behavior of the App using the
API

This commit contributes to the fix of this situation by passing the
wakelock tag in the release method, to enable the release of the proper
wakelock.

Bug: 70691435
Test: ./frameworks/opt/net/wifi/tests/wifitests/runtests.sh

Change-Id: I4e4efd2ce75e77cde54ba92217364b83f2c7c401

Former-commit-id: 3793c020b475fa0728a3035d3790884f128b1ebe