Introduce getProfileSwitchingIcon/Label In CrossProfileApps
Leave the resource retrieval part in client side to avoid passing
drawable across process boundary which should be bad for performance.

The verification part is done in client side, so in theory, apps can work
around the verification. That said, the strings and drawables
are not secret anyway, they are just a generic icon / label to represent
the user. Also, they can always retrieve them by using getDrawable
anyway.

BUG: 67765768
Test: bit FrameworksCoreTests:android.content.pm.crossprofile.CrossProfileAppsTest

Change-Id: I44a7ba17ecc81d0d366184b2d411caa7a82de3ef

Former-commit-id: 8246513dfc5556a76e7db8732d2213919cd27748