AnomalyDetection invalidity depends on MetricProd
DurationMetricProducer has extra conditions that must be satisfied for
an anomaly detection Alert (in the config) to be valid. This test must
only be run for DurationMetricProducer, and not other MetricProducers.

In the future, AnomalyTracker will be split into two, with a
DuratoinAnomalyTracker subclass that will be created by
DurationMetricProducer's createAnomalyTracker function.

Test: CTS test in progress
Change-Id: If172a71d15d5dabb2827859ddb2b5ff7ed386501

Former-commit-id: 450099db169d386077584d93486908c3bae1244e