Fix support for negatively Z-ordered SubWindows.
It turns out the live wallpaper picker is a copy-pasted
implementation of the old version of SurfaceView which relies
on TYPE_APPLICATION_MEDIA to be placed beneath the main window.
This is the last user and we should aim to eliminate it by replacing
the wallpaper picker with a SurfaceView or simply a Child Surface.

Bug: 69928336
Bug: 70040778
Test: Manual. ZOrderingTests.
Change-Id: I462b8f11954ad2e892d23164f41359a73078c21f

Former-commit-id: 0eff18732c7d3c3c7c998737f92da86d479486e9