Provide start parameters via setters on ActivityStarter.
A number of arguments are sent to ActivityStarter when starting an
activity. Not all parameters are relevant to each invocation and many
are later modified as the starter prepares to start the activity.
This can lead to callers providing default values for fields they do
not care about.

This change now sets parameters individually on ActivityStarter,
since it is a single use item. Differing behavior, such as waiting
for the start result, are captured in this manner. Aditionally,
clients also now call execute to start the activity.

Bug: 64750076
Test: atest FrameworksServicesTests:ActivityStartControllerTests
Test: atest FrameworksServicesTests:ActivityStarterTests
Change-Id: Ifc88ae6547bc9155527c31709bef1784c6406bda

Former-commit-id: 4c9a597fc10165b41cb1ffcd575ea800c23d6448