Fix droiddoc when OUT_DIR is an absolute path
The droiddoc rules were passing the sources in through
LOCAL_SRC_FILES, which gets prefixed with $(LOCAL_PATH).  They
were prefixing ../.. to try to counter frameworks/base in
$(LOCAL_PATH), but that doesn't work for absolute paths to
generated files in an absolute $(OUT_DIR).

Use $(LOCAL_GENERATED_SOURCES) for anything that is in $(OUT_DIR)
instead, which takes paths that are relative to the top of the tree
or absolute.

Bug: 70166718
Test: m OUT_DIR=/tmp/out docs
Change-Id: I93b743607c2b8ce05512ecd4bbe860747cb48428

Former-commit-id: bd200ce7d4ebe31ad4327683e59af269420b0ef5