Don't use path to find protoc-gen-javastream
Previous attempts to use a full path in the --plugin argument
failed, but that was because the --plugin argument had the wrong
name (protoc-gen-java-stream instead of protoc-gen-javastream)
and was being completely ignored.  Passing a full path where
the last element of the path is protoc-gen-javastream works fine.

Fixes issues where the command line is expanded with $PATH in
it 58 times which could hit command length limits if $PATH was
even a little bit larger than the default.

Test: m framework
Change-Id: Icd336f522fb8770d9058d18d8a04d24a96405615

Former-commit-id: 6471db9b718b1aeddd861935c1c7780b0c0842c3