Ensure activity stacks have unique ids.
Currently ActivityStackSupervisor will return the first id that it
discovers that is not added to the display. However, stack addition
logic is written in a way where multiple stacks can be created before
being added. This leads to multiple stacks with the same id.

This changelist corrects this collision by always incrementing the
current stack after returning it to the caller. The check to see if the
id is in use has been removed and the bookkeeping is now in
ActivityDisplay.

Change-Id: I2c1b86ef8d74b9aa2eac134ac546c0111fa5e22c
Fixes: 68952156
Test: atest CtsActivityManagerDeviceTestCases:android.server.am.ActivityManagerAppConfigurationTests#testSameConfigurationSplitFullSplitNoRelaunch
Test: atest CtsActivityManagerDeviceTestCases:android.server.am.ActivityManagerAppConfigurationTests#testSameConfigurationSplitFullSplitRelaunch

Former-commit-id: d9cce2cadb791d4d8e913eabb3238268b7c9813f