Restore old stack Z ordering promotion.
This is a more direct translation of what was in WindowLayersController.
We hoped that it wasn't necessary and that they were always in the correct
position in the WindowContainer hierarchy. However it seems there is no mechanism
to position them as such. It's not obvious for me how to resolve things from that angle
so I'm proposing reintroducing this logic.

Bug: 69553456
Bug: 70178829
Test: Manual. go/wm-smoke
Change-Id: I32ab03b1bcd6c498bf643c0f207d3f72286f258b

Former-commit-id: d92e7eb2e9a72ac73e992cf502a080ad47a962f3