Use stable rect to determine if we should adjust for IME in split-screen
Regression was introduced in ag/3174240 where the content rect used to
be cached before the layout IME so the IME didn't affect the content
rect. However, we don't want content rect to mean something differect
here as content rect excludes the IME space in the rest of the system.
So, changed the one use case we have for this to use stable rect.

Change-Id: I13e23baf87dbcaf21a8e0624f73b9af6116f7262
Fixes: 69425766
Test: Steps from bug

Former-commit-id: db5061916b544263a339912b7ca4a1332e9de069