Remove non-junit android.test.runner classes from android.jar
This removes those android.test.** classes which are part of the
android.test.runner stubs jar from the main android.jar.

The junit classes will be removed at a later date. It cannot be done at
the moment because it requires a lot of changes to the build files.

This change is being done now rather than wait until they can all be
removed together because the code base is a moving target. Doing this
change minimizes the window during which other developers could either
revert the previous changes that make this possible or add new code that
relies on the android.test.runner classes being in android.jar and so
would require yet more changes.

Bug: 30188076
Test: make checkbuild
Change-Id: I4e3c3bfea77ce15ca7a9ba7d36aee5396f67affb

Former-commit-id: 1895f2e5bb3c07b2dc85ed6267848e820d6c39b0