Keyguard can be on even if not showing
It is possible for keyguard to be in "restricted input" state even if
isKeyguardShowingAndNotOccluded returns false. The previous CL to
deprecate inKeyguardRestrictedInputMode has also changed the behaviour
of isKeyguardLocked. Return to the old behaviour here, but keep the API
deprecated.

Bug: 70411251
Test: 1) add Google account to phone 2) Enable google assistant and set
up active edge to trigger the assistant 3) Ensure phone is in the
always-on-display mode 4) Squeeze the phone to trigger the assistant.
Previously, the assistant would not launch. The expected behaviour is
that assistant gets launched. Without this change, display will turn on
but the assistant will not get launched.

Change-Id: Ic299f3fc223ebdc3b0f6eea97cb4a89a3142e44a
(cherry picked from commit d57f0d864e178dc84c81448a6cd477ddef6d0855 [formerly 35fbb312e9a185d3d6ed20697b486faae8409463])

Former-commit-id: 5cc27ac6c4a36e60af84e33068f329838aa3858c