Add platform key generation ID to WrappedKey instances
This is so that when we persist them, we can tell that they were wrapped
with a specific version of the platform key. This will be useful for us
to provide error messages to the users of recoverable keys. (i.e., in
the case where the user had an application key that was wrapped with a
platform key that is no longer valid, they MUST rotate key.)

Test: adb shell am instrument -w -e package com.android.server.locksettings.recoverablekeystore com.android.frameworks.servicestests/android.support.test.runner.AndroidJUnitRunner
Change-Id: I91569bcaf23b49d89a9caa9d313d9c93952b620d

Former-commit-id: 9fa18c621e82d4a6e2b647fc3268ddc89e64b73c