Split query permission from update permission
Split time zone rules query permission from update permission.
This would allow an app like Settings to determine the current rules
update state, which would enable it to show the version of rules
the device came with, if there have been any updates and whether
a reboot is needed for changes to take effect.

Test: PTS: run -m PtsTimeZoneTestCases
Test: atest FrameworksServicesTests:RulesManagerServiceTest
Bug: 69443060
Change-Id: Ia5d31814d7caee842f5e28870fab61f7c7129386

Former-commit-id: b214bc44f1942b957f60dad75517e07cede77f18