Add MessagingStyle#isGroupConversation().
This will allow apps to mark a Notification's MessagingStyle as a group
conversation. This is useful because the current implementation
overloads MessagingStyle#conversationTitle; which, when not null,
implies a group conversation. Instead of relying on this implicit
contract, we include this change's explicit field.

Test: runtest -x
core/tests/coretests/src/android/app/NotificationTest.java
Bug: 67717177
Bug: 70725009

Change-Id: I5405b6b0ca9723a06113482c04b8c2f8e1a4318c

Former-commit-id: 9ac617c5a2c53bb04c0425bef7c9ed0f56abf5f3