[Magnifier] Comparing the correct coordinates to in Magnifier update.
Previously we are comparing the clamped coordinates to avoid
unnecessary update of magnifier content. However if the magnifier is
reaching to the edge of the screen it doesn't update itself because
coordinates passed to Magnifier#show() later are clamped to the same.

This CL compares the coordinates without clamping to fix the issue.

Bug: 70580848
Test: bit FrameworksCoreTests:android.widget.TextViewActivityTest
Test: manual test with WebView that shows the magnifier works.
Change-Id: I65d681da334d5037a924e733d1a27cfd77231726

Former-commit-id: 29d1e9eed1a4bedfe883c2009afaa21f91e1f10f