Modify android.test.runner to use new mock APIs
Previous change added some new public APIs to allow the
android.test.runner classes to be built against the public API stubs
rather than the internal classes. This change updates the
android.test.runner classes to use that new API.

This in turn is preparation for building an android.test.legacy library
that can be safely statically included in APKs.

Bug: 30188076
Test: make checkbuild
Change-Id: I302626abb7e048ef0c3169e9a5209b85f2f87d34

Former-commit-id: 20af1df63f77f6b3ab12fd6221f581fe84182091