Remove junit classes from legacy-android-test
A mistake during the conversion of the legacy-android-test target from
Android.mk to Android.bp caused some junit classes to be accidentally
included in the legacy-android-test target. That caused some warnings
about duplicate classes to be output during the build.

The junit classes are being included as part of the static library
android.test.runner. This fixes that by adding a new
android.test.runner-minus-junit which excludes the junit classes
and then using that instead.

Test: make checkbuild
Change-Id: Icd00b38c64d3416a55c1055f9716ddb78ec278f8

Former-commit-id: a88eb29f9e0bd363fb3f1e02f379a54324a3e5ec