Add unique id for sms APIs to be able to trigger correct smstracker
In old APIs the SmsTracker itself was being passed to the RIL
and onSendComplete the tracker's pending intent was triggered
and then the updated messageRef was used as the id.

Instead of passing the tracker, passing a unique id to be used
for the lifetime of the message.

Doing same thing for receiving flow.

Test: None, APIs not exercised yet.
BUG=69846044
Merged-In: Id19f854e2c48649db8f2031ee4f49cdac331451c
Change-Id: Id19f854e2c48649db8f2031ee4f49cdac331451c

Former-commit-id: 967a952b34d08427f0aeb69f1328417b8ad7ffc3