Switch to SkAndroidCodec::computeSampleSize
Bug: 63909536
Test: CtsGraphicsTestCases(ImageDecoderTest) and Skia

Depends on https://skia-review.googlesource.com/94620 in Skia.

Move the code for determining the actual sample size to Skia.
That code includes tests which caught bugs that are hard to
test for from the Java API. (Previously, we could end up using
the wrong sample size and then scaling unnecessarily.)

Change-Id: Ib3a5d70e4845e95b18af77668383d38746cb21fc

Former-commit-id: ea978db1f86b8c6fae3683eafe6621fd958308d7