Only have one way of dumping print manager state
Add a class DualDumpOutputStream that maps proto-dump commands onto
print writer commands.

The effect of this is that there is only one - very proto dump like -
way to dump the print manager which is much easier to maintain.

The DualDumpOutputStream tries to produce a result similar to the
incident-report tool.

Test: adb shell dumpsys print
Change-Id: I1f0c56651eaa59f0ce90cdb08c71e89a96c48dd4

Former-commit-id: 9a534c01ecd4a1442856c1b709e23b61f1baf9ba