Fixes dock side when rotating to seascape after drag to split
Entering split screen (in portrait or landscape) using drag to top
and then rotate to seascape docks primary stack to wrong side. When
entering with drag, display content creates a stack and tries to query
dock side without setting itself into the stack which returns an invalid
dock side. The dock side is used by DockedStackDividerController to hold
the original side that the primary stack split to when
notifyDockedStackExistsChanged occurs. Fix by defering calling
notifyDockedStackExistsChanged until display content is added to the
stack.

Test: manual - drag app in overview to split and rotate CCW
Change-Id: I14f5d68ab95548ee7276c7101c5f991a5ceb69e4
Fixes: 70718233

Former-commit-id: e8b052af0827a5d84b68f55f4b9f86e57e7ca70e