Fix crash in EditText#getText
The fact that View's constructor may call other methods before the
TextView constructor has a chance to initialize values means that
getText() may return null in the rare case where it is called
before the constructor.

Protect getText() from running into errors when the text is null,
as it just means the value is not ready yet.

Bug: 71638589
Test: CTS in topic
Change-Id: Ie11233416385353f78e3ee27fb0d3ec9b671512f

Former-commit-id: b102cc24354fb2e44964b60a794746df86be1eb7