Switch activity relaunch to lifecycler
This is needed to be able to relaunch an activity and eventually
bring it to stopped state while it is not visible. Before this
relaunch request put activity in resumed state in the end. With
lifecycler both relaunch and request to move to stopped state can
be done in a single transaction to client.

Also fixed extra relaunch request to activity which was sent
before it was even started. It happened because we were waiting
for another activity to pause before actually scheduling the
launch, and in activity-paused callback we were ensuring config
first. This led to configuration update for not-yet launched
activity and relaunch message was sent first.

Bug: 64610483
Test: android.app.servertransaction tests
Change-Id: Ib89c20cd0fdff209319d4819023c3cdd641a72c7

Former-commit-id: b372da6ef81cd56f681f95d042692922b72b5751