WebView: document Safe Browsing as 'on' by default
WebView Safe Browsing will be on by default. While previously it was
opt-in (applications needed to enable via manifest or API), now it is an
opt-out feature.

This CL updates documentation to mention that the feature is on by
default and provides steps for applications to opt-out if necessary.

This also makes the case of "Safe Browsing" consistent across
documentation.

Bug: 69579509
Test: make docs (manually verify)
Change-Id: I77683a684cd79d2e3cd313ddbff8c52fe510b72b

Former-commit-id: fb92ee13486acab7acbd68a0f17024af065590a3