Uses back and home button for screen pinning when recents is invisible
All the text and screen pinning hint shows tells the user that the back
and home button should be held to exit screen pinning. The hint also do
not have the recents button shown if it is invisible. The toast code has
been moved to recents and services calls through to post a toast message
depending if the recents button is visible.

Test: manual
Fixes: 72059911
Change-Id: I93abf5072b97760f33e7e77421544a4b3ad27beb

Former-commit-id: c4356fb75318967148e8394878af1c13d37080dd