Exempting allow-while-idle alarms in EBS
Exempting allow-while-idle alarms when the device is in Extreme battery
saver. The throttling of one allow-while-idle alarm per
ALLOW_WHILE_IDLE_LONG_TIME is applied even when the device may not be
dozing. The throttling is relaxed to ALLOW_WHILE_IDLE_SHORT_TIME if the
uid was in the foreground recently.

Test: atest android.alarmmanager.cts.BatterySaverTests

Bug: 72124522
Change-Id: Ic75bac6234745c73a345180387c83bc725823aa8

Former-commit-id: c25447daa1ce6b51a8fe8eb13b376220c62849bd