Do not propagate parent bound changes from task as onResize.
When a stack is resized on the WindowManager side, it does not
itself signal a resize in any children. The children themselves will
receive updated override configurations for this. It is therefore not
necesssary to propagate calls from the task to the children when the
stack is resized.

This changelist separates out the callback for parent resizing into
onParentResize. The default implementation is to call onResize. The
default implementation of onResize now calls onParentResize on its
children. For Task, onParentResize is overridden to prevent
propagation.

Change-Id: If7990c88f12245774d085650b15da9abc96bf622
Fixes: 70558903
Test: atest FrameworksServicesTests:com.android.server.wm.WindowContainerTests#testOnParentResizePropagation

Former-commit-id: d92ae48606fd89f0fffd570db8d3175cd6f82728