Let the transport ask for non-incremental backup.
Add a new performBackup() return code TRANSPORT_PACKAGE_DATA_LOST. If
the transport returns this then delete the state for the package and
immediately retry it. Thus the retry will be a non-incremental backup.

I have left the new constant as @hide to unblock this change. I will add
this to the current API review change: ag/3477872.

Bug: 71887023
Test: I tested manually by returning the new code for both @pm@ and
other key value packages. No automated test exists for this file, and it
will be tricky to add this before IC.

Change-Id: I4357a8b297c373eb81e25afb82d480b26f6642b6

Former-commit-id: f5b99ad2ea12e65ac66321fdf57e3b031f3985d0