Introduce InputConnection#reportLanguageHint()
The goal of this CL is to introduce a way for IMEs to tell
applications about what languages are likely to be used in the text to
be committed, for IMEs that support dynamic language switching
(especially without relying on InputMethodSubtype).

It is already obvious that InputMethodSubtype is not working well for
this use case.  We need is a simple, easy, and back-portable way that
can be widely accepted by the ecosystem.

Bug: 7031513
Test: atest CtsInputMethodTestCases
Change-Id: Ie86edafd1ed68b58f702116f561fc448fdbb57a8

Former-commit-id: e77386e8fbe01b7fbf1a04555d4088796d862f03