Revert "Use ImageDecoder for BitmapDrawable"
Bug: 72381918
Test: Covered by existing tests

This CL seems to have broken the theme tests. In addition, it changes
the behavior of a couple of cases of passing InputStreams to the
framework. Previously, the framework used BitmapFactory, which did not
close the InputStreams, but ImageDecoder does.

Planning to reland along with the fix for closing in ag/3497523 and the
fix for the theme tests.

This reverts commit 1cf5b5c88d4636326b829491f10be2c6f26d2389 [formerly 66c6d78908e0203a497c2e97398c824b2591ea19].

Change-Id: I4ac6d0f2e1e2bb0925ae71c141bfe8a0d37e6a16

Former-commit-id: e7addfc99909b8e304f5e3ec3bda4cd938197fc2