Use config_headlineFontFamilyMedium for actionBar/toolbar.
Updates the fontFamily used for Theme.DeviceDefault.Settings's
actionBarStyle's titleTextStyle to be config_headlineFontFamilyMedium.

Also updates the toolbarStyle for DeviceDefault themes to use a new
DeviceDefault version that sets titleTextAppearance to a DeviceDefault
version that picks up the fontFamily change made above.

This allows greater customization of system surfaces for OEMs.

Bug: 69314526
Test: visual inspection
Change-Id: Ida9bfaba1e136f29ae80715d1648f13f7b2f2126

Former-commit-id: 569bfdda6c17440dff38b37a5cba89815cd00739