DND qs tile primary text is always Do not disturb
In P, do not disturb in quick settings always uses priority-only
do not disturb, so there is no reason to distinguish between
priority-only/alarms only/total silence modes.  Users can still change
the behavior of priority-dnd in settings, but this will not chang the
primary text of the quick settings tile for dnd.

Change-Id: I5b6e911cb2542a7a1d0696871607eb8abc95cd43
Fixes: 72418705
Test: manual (toggle on/off dnd qs tile)

Former-commit-id: 7999ba90c6bb692f68b8b733114a4e786bc6623f