Prevent writeDataToDisk calls when switching from guest user to primary user
Previsouly, we were only listening to ACTION_SHUTDOWN event, which is
triggered whenever 'user' on Android is changed. This will fire an
intent to write a file to disk and it is not intended. So we add
an additional check to see if the ACTION_SHUTDOWN is on foreground as
user switch fires ACTION_SHUTDOWN is NOT on foreground.

Bug: 72461610
Bug: 72441801
Test: manual testing
Change-Id: I05ea8f1b60b6644d1e6ccf232d6dcde39538666d

Former-commit-id: f7a3bcbeb7b89d35a916ded587bee6012061c1ff