Send broadcast to pkg installers on config change
Send a protected system broadcast to apps that hold
the INSTALL_PACKAGES permission whenever a change to
properties like locale or display density happens.

Receivers can be registered in the manifest, and a process
will be woken up to receive the broadcast.

This will happen rarely, since the configuration changes
eligible to trigger this broadcast are rare themselves.

Bug: 63918966
Test: manual

Change-Id: I817a51ea05f762e02561691825d57d643db7dc30

Former-commit-id: 57fccd61a53b962f568235dccf9e8bc23005a41b