Reset brightness adjustment when user sets brightness directly.
If the user has a non-zero brightness adjustment then short-term model
returns an incorrect value after they make a direct change to the
brightness with auto-brightness on, which then looks like the
auto-brightness is ignoring their input. To avoid this , if they're
making a direct change to auto-brightness go ahead and reset the
brightness adjustment.

Change-Id: I257a6ecc7ca3df0e610710df19d81a65b009af45
Fixes: 72506596
Test: manual

Former-commit-id: 617564ff2cdee2e9a50339544129283503ae71a8