Clear pending brightness setting when it's redundant.
If we don't clear the pending brightness setting when it's equivalent to
the existing brightness setting then we end up racing the next time
auto-brightness changes. Most of the time the settings change
notification will win and we'll ignore the stale pending brightness, but
sometimes we immediately update the power state again and then we'll see
the stale value which looks different from our current value and so
assume it's a change coming from the user. Whoops.

Bug: 72508527
Test: manual
Change-Id: Id743f3654be6d48a14fca5882c1b8e16fed147da

Former-commit-id: 44dc8b1da8ec60095524c391fd9b8f1c907e76ea