Always execute onStop and onSaveInstanceState together
Because onStop is guaranteed to be called when activity goes away
and onSaveInstanceState is always called before onStop (starting
with Honeycomb), we can execute these two callbacks together.

This CL consolidates all call points to stop the activity and
moves them to callActivityOnStop, which will also take care of
saving instance state.

Bug: 68325009
Bug: 68763258
Test: Manual
Change-Id: I1062c6be28581a7de7961babd55664640aaae359

Former-commit-id: 8ae795739070c4ebbde840d25845f3ddde2a606a