Revert "Revert "Adds generic intent Instant App resolution""
This reverts commit 5401a5a54ebc2d1448d4b4e28395617d6cc7b9da [formerly 860b8ba71938e9860a31881c1d1431877f9d01a2].

The original change that was reverted contained a bug that allowed an
http view/browsable intent used to query for browsers to be considered
as a candidate for instant apps. This was resulting in an attempt to
bind to the instant app resolver while holding a lock on mPackages.

This change ensures that PMS doesn't bind while checking for the browser
status of a package in both the instant app filtering code and by adding
the FLAG_IGNORE_EPHEMERAL to the canonical browser intent.

Reason for revert: Applying fix

Change-Id: I4896b3a15416a11fdc3f6c191e552c4ce8963623
Fixes: 63117034
Fixes: 71916178
Test: Manual using test app at google_experimental/users/patb/InstantAppsInP
Test: atest android.appsecurity.cts.EphemeralTest passes after modification

Former-commit-id: 577d402d0d938c14d415054289a5ecbc613d0046